change database schema
The goal here was to change the database to better reflect the game
state and allow for better searching of data. It was more or less not
possible to represent some things with the previous database structure.
Most notably, relationships that existed for no purpose have been
removed, additional info for game state tracking has been added.