Add populate DB info
node populateDB will allow you to put 1000 (default) rows of random
values into the database, as of the createDB in this script. It should
work on any database so long as entries haven't been deleted yet. The
other changes are incindental, and were because of a small start to the
vehicle js. I'll switch to the vehicles branch to finish those.